결제 기능 및 테스트 추가

+++ src/main/java/com/example/hhp_ecommerce/interfaces/dto/payment/PaymentRequestDto.java
@@ -1,4 +1,4 @@
 package com.example.hhp_ecommerce.interfaces.dto.payment;
 
-public record PaymentRequestDto(long orderId, long userId, long amount) {
+public record PaymentRequestDto(long orderId, long userId, int amount) {
 }
